+++ src/accelerate/utils/__init__.py
@@ -63,12 +63,14 @@ from .environment import (
     are_libraries_initialized,
     check_cuda_p2p_ib_support,
     check_fp8_capability,
+    clear_environment,
     convert_dict_to_env_variables,
     get_cpu_distributed_information,
     get_gpu_info,
     get_int_from_env,
     parse_choice_from_env,
     parse_flag_from_env,
+    patch_environment,
     set_numa_affinity,
     str_to_bool,
 )
@@ -245,14 +247,12 @@ from .memory import find_executable_batch_size, release_memory
 from .other import (
     check_os_kernel,
     clean_state_dict_for_safetensors,
-    clear_environment,
     convert_bytes,
     extract_model_from_parallel,
     get_pretty_name,
     is_port_in_use,
     load,
     merge_dicts,
-    patch_environment,
     recursive_getattr,
     save,
     wait_for_everyone,
